backend/requirements.txt
Flask
openai
python-dotenv
chromadb
pypdf # For PDF parsing
langchain
